requirements.txt
# Identidad y Configuración
python-dotenv>=1.0.0

# El conector oficial para SQL Warehouses
databricks-sql-connector>=3.0.0
databricks-sqlalchemy>=2.0.9

# Kedro y Conectores de datos
kedro~=1.2.0
kedro-datasets[pandas,sql]>=9.1

# Desarrollo
ipython>=8.10